+++ test/support/mocks.ex
@@ -1,3 +1,4 @@
 Mox.defmock(ExQueueBusClient.EventHandlerMock, for: ExQueueBusClient.EventHandlerBehaviour)
 Mox.defmock(ExQueueBusClient.SNSClientMock, for: ExQueueBusClient.SNSClientBehaviour)
 Mox.defmock(ExQueueBusClient.AWSClientMock, for: ExQueueBusClient.AWSClientBehaviour)
+Mox.defmock(ExQueueBusClient.ConfigMock, for: ExQueueBusClient.BusSupervisor.Config)